fix(tui): parse min_duration as float in trace view
Change int() to float() when parsing the #trace-min-duration input so
values like '2.5' are accepted and forwarded to the agent without
truncation.  The default when the field is empty changes from 0 to 0.0
to maintain float type consistency.  Existing negative-value rejection
and ValueError handling are preserved.

Tests: TestTraceView
  - test_float_min_duration_accepted
  - test_negative_float_min_duration_rejected

+++ peeka/tui/views/trace.py
@@ -617,7 +617,7 @@ class TraceView(Container):
 
         # Parse min_duration, default to 0
         try:
-            min_duration_ms = int(min_duration_input) if min_duration_input.strip() else 0
+            min_duration_ms = float(min_duration_input) if min_duration_input.strip() else 0.0
             if min_duration_ms < 0:
                 self.app.notify("Min duration must be >= 0", severity="warning")
                 return
